requirements.txt
agora_token_builder==1.0.0
altgraph==0.17.4
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
cachetools==5.3.2
celery==5.3.4
certifi==2021.10.8
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.36.0
colorama==0.4.6
constantly==15.1.0
distlib==0.3.6
Django==3.2.15
django-autoslug==1.9.9
django-ckeditor==6.7.1
django-cloudinary-storage==0.3.0
django-environ==0.10.0
django-js-asset==2.2.0
django-model-utils==4.3.1
django-notifications-hq==1.7.0
filelock==3.12.0
geographiclib==2.0
idna==3.4
image==1.5.33
incremental==21.3.0
jsonfield==3.1.0
kombu==5.3.2
msgpack==1.0.4
numpy==1.26.0
opencv-python==4.9.0.80
packaging==23.2
pefile==2023.2.7
pillow==10.2.0
platformdirs==3.5.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.3
py2exe==0.13.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
PyJWT==2.6.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
pywin32-ctypes==0.2.2
requests==2.31.0
sentry-sdk==1.5.12
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
tenant-schemas-celery==2.2.0
twisted-iocpsupport==1.0.4
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.7
urllib3==1.26.9
vine==5.0.0
virtualenv==20.23.0
wcwidth==0.2.5
whitenoise==6.1.0
wrapt==1.14.1
xlwt==1.3.0
zope.interface==5.4.0
